fix(field): 调整 field color 为 antd 新的 ColorPicker
