+++ packages/field/src/components/ColorPicker/index.tsx
@@ -1,154 +1,81 @@
-﻿import { proTheme, useStyle } from '@ant-design/pro-provider';
-import type { SketchPickerProps } from '@chenshuai2144/sketch-color';
-import { SketchPicker } from '@chenshuai2144/sketch-color';
-import type { PopoverProps } from 'antd';
-import { ConfigProvider, Popover } from 'antd';
-
-import useMergedState from 'rc-util/lib/hooks/useMergedState';
-import React, { useContext, useImperativeHandle } from 'react';
+﻿import { ColorPicker } from 'antd';
+import type { PresetsItem } from 'antd/es/color-picker/interface';
+import React, { useImperativeHandle, useRef } from 'react';
 import type { ProFieldFC } from '../../index';
-
-export const DEFAULT_COLORS = [
-  '#FF9D4E', // 0 - 橘黄色
-  '#5BD8A6', // 1 - 绿色
-  '#5B8FF9', // 2 - 蓝色
-  '#F7664E', // 3 - 红色
-  '#FF86B7', // 4 - 水红色
-  '#2B9E9D', // 5 - 墨绿色
-  '#9270CA', // 6 - 紫色
-  '#6DC8EC', // 7 - 浅蓝色
-  '#667796', // 8 - 黛蓝色
-  '#F6BD16', // 9 - 黄色
-];
-
-const ColorPicker = React.forwardRef(
-  (
-    {
-      mode,
-      popoverProps,
-      ...rest
-    }: SketchPickerProps & {
-      value?: string;
-      popoverProps?: PopoverProps;
-      mode?: 'read' | 'edit';
-      onChange?: (color: string) => void;
-      colors?: string[];
-      disabled?: boolean;
-    },
-    ref,
-  ) => {
-    const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
-    const prefixCls = getPrefixCls('pro-field-color-picker');
-
-    const { token } = proTheme.useToken();
-    const [color, setColor] = useMergedState('#1890ff', {
-      value: rest.value,
-      onChange: rest.onChange,
-    });
-
-    const { wrapSSR, hashId } = useStyle('ProFiledColorPicker' + color, () => {
-      return {
-        [`.${prefixCls}`]: {
-          width: 32,
-          height: 32,
-          display: 'flex',
-          alignItems: 'center',
-          justifyContent: 'center',
-          boxSizing: 'border-box',
-          border: `1px solid ${token.colorSplit}`,
-          borderRadius: token.borderRadius,
-          '&:hover': {
-            borderColor: color,
-          },
-        },
-      };
-    });
-
-    const readDom = wrapSSR(
-      <div
-        className={`${prefixCls} ${hashId}`.trim()}
-        style={{
-          cursor: rest.disabled ? 'not-allowed' : 'pointer',
-          backgroundColor: rest.disabled
-            ? token.colorBgContainerDisabled
-            : token.colorBgContainer,
-        }}
-      >
-        <div
-          style={{
-            backgroundColor: color,
-            width: 24,
-            boxSizing: 'border-box',
-            height: 24,
-            borderRadius: token.borderRadius,
-          }}
-        />
-      </div>,
-    );
-
-    useImperativeHandle(ref, () => {});
-
-    if (mode === 'read' || rest.disabled) {
-      return readDom;
-    }
-
-    return (
-      <Popover
-        trigger="click"
-        placement="right"
-        {...popoverProps}
-        content={
-          <div
-            style={{
-              margin: '-12px -16px',
-            }}
-          >
-            <SketchPicker
-              {...rest}
-              presetColors={rest.colors || rest.presetColors || DEFAULT_COLORS}
-              color={color}
-              onChange={({ hex, rgb: { r, g, b, a } }) => {
-                if (a && a < 1) {
-                  setColor(`rgba(${r}, ${g}, ${b}, ${a})`);
-                  return;
-                }
-                setColor(hex);
-              }}
-            />
-          </div>
-        }
-      >
-        {readDom}
-      </Popover>
-    );
-  },
-);
-
-/**
- * 颜色组件
- *
- * @param FieldColorPicker {
- *     text: number;
- *     moneySymbol?: string; }
- */
-const FieldColorPicker: ProFieldFC<{
+// https://ant.design/components/color-picker-cn 示例颜色
+const DEFAULT_PRESETS: PresetsItem = {
+  label: 'Recommended',
+  colors: [
+    // '#000000',
+    // '#000000E0',
+    // '#000000A6',
+    // '#00000073',
+    // '#00000040',
+    // '#00000026',
+    // '#0000001A',
+    // '#00000012',
+    // '#0000000A',
+    // '#00000005',
+    '#F5222D',
+    '#FA8C16',
+    '#FADB14',
+    '#8BBB11',
+    '#52C41A',
+    '#13A8A8',
+    '#1677FF',
+    '#2F54EB',
+    '#722ED1',
+    '#EB2F96',
+    '#F5222D4D',
+    '#FA8C164D',
+    '#FADB144D',
+    '#8BBB114D',
+    '#52C41A4D',
+    '#13A8A84D',
+    '#1677FF4D',
+    '#2F54EB4D',
+    '#722ED14D',
+    '#EB2F964D',
+  ],
+}
+const FieldColor: ProFieldFC<{
   text: string;
-}> = ({ text, mode: type, render, renderFormItem, fieldProps }, ref: any) => {
-  if (type === 'read') {
-    const dom = <ColorPicker value={text} mode="read" ref={ref} />;
-    if (render) {
-      return render(text, { mode: type, ...fieldProps }, dom);
+}> = (
+  { text, mode, render, renderFormItem, fieldProps },
+  ref,
+) => {
+    const inputRef = useRef<HTMLInputElement>();
+    useImperativeHandle(ref, () => inputRef.current, []);
+    if (mode === 'read') {
+      const dom = (
+        <ColorPicker
+          ref={inputRef}
+          presets={[DEFAULT_PRESETS]}
+          {...fieldProps}
+          // 设置无法 open 
+          open={false}
+        />
+      );
+      if (render) {
+        return render(text, { mode, ...fieldProps }, dom);
+      }
+      return dom;
     }
-    return dom;
-  }
-  if (type === 'edit' || type === 'update') {
-    const dom = <ColorPicker ref={ref} {...fieldProps} />;
-    if (renderFormItem) {
-      return renderFormItem(text, { mode: type, ...fieldProps }, dom);
-    }
-    return dom;
-  }
-  return null;
-};
+    if (mode === 'edit' || mode === 'update') {
+      const dom = (
+        <ColorPicker
+          ref={inputRef}
+          presets={[DEFAULT_PRESETS]}
+          {...fieldProps}
+        />
+      );
 
-export default React.forwardRef(FieldColorPicker);
+      if (renderFormItem) {
+        return renderFormItem(text, { mode, ...fieldProps }, dom);
+      }
+      return dom;
+    }
+    return null;
+  };
+
+export default React.forwardRef(FieldColor);
